@@ -1,8 +1,12 @@
 <?php
 
+namespace Database\Seeders;
+
 use Illuminate\Database\Seeder;
 use Illuminate\Support\Facades\DB;
 use Faker\Factory as Faker;
+use App\Models\User;
+use App\Models\Job;
 
 class JobsTableSeeder extends Seeder
 {
@@ -14,11 +18,11 @@ class JobsTableSeeder extends Seeder
     public function run()
     {
         // create 10 users, for each user create and asssociate 10 jobs with that user
-        factory(App\User::class, 10)->create()->each(function ($user)
+        User::factory()->count(10)->create()->each(function ($user)
         {
             $user->jobs()
                 ->saveMany(
-                    factory(App\Job::class, 10)->make());
+                    Job::factory()->count(10)->make());
         });
     }
 }
