@@ -33,7 +33,7 @@ class RegistrationRequest extends FormRequest
         return [
             // validation rules
             'name' => ['required', 'string', 'max:255'],
-            'email' => ['required','email','unique:App\User,email'],
+            'email' => ['required','email','unique:App\Models\User,email'],
             'password' => ['required']
         ];
     }
