@@ -1,11 +1,14 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 
 class ApplicationNote extends Model
 {
+    use HasFactory;
+    
     /**
      * The table associated with the model.
      *
