@@ -1,11 +1,14 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 
 class Interview extends Model
 {
+    use HasFactory;
+    
     /**
      * The attributes that are mass assignable.
      *
